+++ alicompat/neoforge/src/compat/farlanders/java/com/yanny/alicompat/compat/farlanders/FarlandersCompat.java
@@ -1,6 +1,7 @@
 package com.yanny.alicompat.compat.farlanders;
 
 import com.yanny.ali.api.IServerRegistry;
+import com.yanny.ali.api.TradeLevel;
 import com.yanny.alicompat.IModCompat;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
@@ -30,7 +31,7 @@ public class FarlandersCompat implements IModCompat {
     private static void registerTrader(IServerRegistry registry, String name, String tradeSetPrefix, int levels) {
         Identifier traderId = id(name);
 
-        registry.registerTrades(traderId, BuiltInRegistries.ENTITY_TYPE.getValue(traderId), () -> tradeSets(tradeSetPrefix, levels));
+        registry.registerTrades(traderId, BuiltInRegistries.ENTITY_TYPE.getValue(traderId), () -> TradeLevel.ofSets(tradeSets(tradeSetPrefix, levels)));
     }
 
     @NotNull