+++ alicompat/neoforge/src/compat/occultism/java/com/yanny/alicompat/compat/occultism/OccultismLang.java
@@ -14,6 +14,24 @@ public class OccultismLang implements ICompatTranslations {
 
     public static final Map<String, String> TRANSLATION_MAP = new HashMap<>();
 
+    public enum Value implements ITooltipKey {
+        WITH_THIRD_EYE("with_third_eye", "Requires Third Eye, Otherworld Goggles or True Sight Staff"),
+        WITHOUT_THIRD_EYE("without_third_eye", "Without Third Eye, Otherworld Goggles or True Sight Staff"),
+        ;
+
+        private final Translation translation;
+
+        Value(String k, String e) {
+            translation = new Translation("alicompat.property.value." + k, e);
+        }
+
+        @NotNull
+        @Override
+        public Translation getTranslation() {
+            return translation;
+        }
+    }
+
     public enum ConsumeEffects implements ITooltipKey {
         DAMAGE_ITEM("damage_item", "Damage Item:"),
         ;
@@ -32,6 +50,7 @@ public class OccultismLang implements ICompatTranslations {
     }
 
     static {
+        CoreLang.register(TRANSLATION_MAP, Value.class);
         CoreLang.register(TRANSLATION_MAP, ConsumeEffects.class);
     }